requirements.txt
PyMySQL==1.0.2

docker~=5.0.3
PyYAML~=5.3.1
elasticsearch~=7.5.1
Flask~=2.0.3
tqdm~=4.63.0
requests~=2.26.0